benchmarks/requirements.txt
upskill>=0.1.0
pyyaml>=6.0
